Fixed full screen test for smaller displays
